// MeskChatClient/src/types/responseEntityOfListOfApplicationUser.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * MeskChatApplication.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ApplicationUser2 } from './applicationUser2';
import type { ResponseEntityOfListOfApplicationUserValidationErrors } from './responseEntityOfListOfApplicationUserValidationErrors';

export interface ResponseEntityOfListOfApplicationUser {
  isSuccess?: boolean;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  data?: ApplicationUser2[] | null;
  statusCode?: number;
  /** @nullable */
  validationErrors?: ResponseEntityOfListOfApplicationUserValidationErrors;
}